+++ commands/Bot/invite.js
@@ -7,12 +7,14 @@ module.exports = {
     description: "Bot's invite link",
     usage: "invite\n**e.g.**\n\`invite\`\n> Returns a link to Invite the bot on another server",
     run: async (client, msg, arg) => {
-        msg.delete();
-        const embed = new Discord.MessageEmbed()
+        if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
+
+        const inviteEmbed = new Discord.MessageEmbed()
             .setColor(`RANDOM`)
             .setThumbnail(client.user.avatarURL)
             .setTitle(`**Chill** BOT`)
             .setDescription(`[**Invite**](${config.bot_invite_link})`);
-        msg.channel.send(embed)
+
+        msg.channel.send(inviteEmbed)
     }
 }
