+++ commands/Fun/8ball.js
@@ -7,8 +7,7 @@ module.exports = {
     description: "Ask a question, the Bot will answer you",
     usage: "8ball <question>\n**e.g.**\n\`8ball is this bot cool?\`\n> For better results you should only ask questions which can be answered with a simple yes or no\n> Based on the popular 1950 Magic 8-Ball game",
     run: async (client, msg, arg) => {
-
-        msg.delete();
+        if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
 
         const noquestionEmbed = new Discord.MessageEmbed()
             .setColor(`RED`)
