+++ commands/Music/nowplaying.js
@@ -7,7 +7,7 @@ module.exports = {
     description: "Return current playing song",
     usage: "nowplaying\n**e.g.**\n\`nowplaying\`\n> Wanna know the name of the song currently playing?\n> This is the right command",
     run: async (client, msg, arg) => {
-        msg.delete();
+        if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
         const serverQueue = client.queue.get(msg.guild.id);
 
         const noDJroleEmbed = new Discord.MessageEmbed()
