+++ commands/Music/volume.js
@@ -7,7 +7,8 @@ module.exports = {
     permission: "ADMINISTRATOR",
     usage: "volume [value]\n**e.g.**\n\`volume\`\n> Get current volume value\n\`volume 10\`\n> Set volume to 10\n> Default value is 5, it is not reccomended to set too high values to avoid audio distortion\n> Values are logarithmic, so a small value increase will provide a big volume boost",
     run: async (client, msg, arg) => {
-        msg.delete();
+        if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
+        
         const serverQueue = client.queue.get(msg.guild.id);
         const newvolume = arg[0];
