+++ commands/Settings/resetconfig.js
@@ -17,7 +17,8 @@ const defaultSettings = {
 	supportrole: "Support",
 	roleonjoin: "Member",
 	musicchannelonly: "false",
-	xpcooldown: 5
+	xpcooldown: 5,
+	autodeletecmds: "true"
 }
 
 module.exports = {
@@ -28,7 +29,7 @@ module.exports = {
 	usage: "resetconfig\n**e.g.**\n\`resetconfig\`\n> reset the server config values back to default",
     permission: "ADMINISTRATOR",
     run: async (client, msg, arg) => {
-		msg.delete();
+		if (client.settings.get(msg.guild.id, "autodeletecmds") === "true") msg.delete();
 
 		const nopermEmbed = new Discord.MessageEmbed()
 		.setColor(`RED`)
